Audit date-time agnostic components
Entire-Checkpoint: 7188e6156b7b
