--- crates/ars-components/tests/proptest_state_machines/date_time.rs
@@ -1,1276 +1,0 @@
-use ars_components::date_time::{
-    calendar::{self, PageBehavior, SelectionMode},
-    date_field::{self, DateSegmentKind},
-    date_picker, date_range_field, date_range_picker, date_time_picker, range_calendar, time_field,
-};
-use ars_core::{ComponentPart, ConnectApi, Env, KeyboardKey, Service};
-use ars_i18n::{CalendarDate, CalendarDateTime, DateRange, HourCycle, Time, Weekday};
-use proptest::prelude::*;
-
-#[derive(Clone, Debug)]
-enum DateFieldAction {
-    Send(date_field::Event),
-    SetControlledValue(Option<CalendarDate>),
-    SetDisabled(bool),
-    SetReadonly(bool),
-    SetInvalid(bool),
-}
-
-#[derive(Clone, Debug)]
-enum TimeFieldAction {
-    Send(time_field::Event),
-    SetControlledValue(Option<Time>),
-    SetDisabled(bool),
-    SetReadonly(bool),
-    SetInvalid(bool),
-}
-
-fn date(year: i32, month: u8, day: u8) -> CalendarDate {
-    CalendarDate::new_gregorian(year, month, day).expect("generated date should be valid")
-}
-
-fn date_field_props() -> date_field::Props {
-    date_field::Props::new().id("date-field").label("Date")
-}
-
-fn time_field_props() -> time_field::Props {
-    time_field::Props::new()
-        .id("time-field")
-        .label("Time")
-        .hour_cycle(Some(HourCycle::H12))
-        .granularity(time_field::TimeGranularity::Second)
-}
-
-fn arb_date() -> impl Strategy<Value = CalendarDate> {
-    (1900i32..=2100, 1u8..=12, 1u8..=28).prop_map(|(year, month, day)| date(year, month, day))
-}
-
-fn arb_time() -> impl Strategy<Value = Time> {
-    (0u8..=23, 0u8..=59, 0u8..=59)
-        .prop_map(|(hour, minute, second)| Time::new(hour, minute, second, 0).unwrap())
-}
-
-fn arb_editable_kind() -> impl Strategy<Value = DateSegmentKind> {
-    prop_oneof![
-        Just(DateSegmentKind::Year),
-        Just(DateSegmentKind::Month),
-        Just(DateSegmentKind::Day),
-    ]
-}
-
-fn arb_digit() -> impl Strategy<Value = char> {
-    (0u8..=9).prop_map(|digit| char::from(b'0' + digit))
-}
-
-fn arb_time_editable_kind() -> impl Strategy<Value = DateSegmentKind> {
-    prop_oneof![
-        Just(DateSegmentKind::Hour),
-        Just(DateSegmentKind::Minute),
-        Just(DateSegmentKind::Second),
-        Just(DateSegmentKind::DayPeriod),
-    ]
-}
-
-fn arb_date_field_action() -> impl Strategy<Value = DateFieldAction> {
-    prop_oneof![
-        arb_editable_kind()
-            .prop_map(|kind| DateFieldAction::Send(date_field::Event::FocusSegment(kind))),
-        Just(DateFieldAction::Send(date_field::Event::BlurAll)),
-        arb_editable_kind()
-            .prop_map(|kind| DateFieldAction::Send(date_field::Event::IncrementSegment(kind))),
-        arb_editable_kind()
-            .prop_map(|kind| DateFieldAction::Send(date_field::Event::DecrementSegment(kind))),
-        (arb_editable_kind(), arb_digit()).prop_map(|(kind, ch)| DateFieldAction::Send(
-            date_field::Event::TypeIntoSegment(kind, ch)
-        )),
-        arb_editable_kind()
-            .prop_map(|kind| DateFieldAction::Send(date_field::Event::TypeBufferCommit(kind))),
-        Just(DateFieldAction::Send(date_field::Event::CompositionStart)),
-        (arb_editable_kind(), "[0-9]{0,4}".prop_map(String::from)).prop_map(|(kind, text)| {
-            DateFieldAction::Send(date_field::Event::CompositionEnd(kind, text))
-        }),
-        arb_editable_kind()
-            .prop_map(|kind| DateFieldAction::Send(date_field::Event::ClearSegment(kind))),
-        Just(DateFieldAction::Send(date_field::Event::ClearAll)),
-        prop::option::of(arb_date())
-            .prop_map(|value| DateFieldAction::Send(date_field::Event::SetValue(value))),
-        prop::option::of(arb_date()).prop_map(DateFieldAction::SetControlledValue),
-        any::<bool>().prop_map(DateFieldAction::SetDisabled),
-        any::<bool>().prop_map(DateFieldAction::SetReadonly),
-        any::<bool>().prop_map(DateFieldAction::SetInvalid),
-    ]
-}
-
-fn arb_time_field_action() -> impl Strategy<Value = TimeFieldAction> {
-    prop_oneof![
-        arb_time_editable_kind()
-            .prop_map(|kind| TimeFieldAction::Send(time_field::Event::FocusSegment { kind })),
-        Just(TimeFieldAction::Send(time_field::Event::BlurAll)),
-        arb_time_editable_kind()
-            .prop_map(|kind| TimeFieldAction::Send(time_field::Event::IncrementSegment { kind })),
-        arb_time_editable_kind()
-            .prop_map(|kind| TimeFieldAction::Send(time_field::Event::DecrementSegment { kind })),
-        (
-            arb_time_editable_kind(),
-            prop_oneof![arb_digit(), Just('a'), Just('p')]
-        )
-            .prop_map(|(kind, ch)| TimeFieldAction::Send(
-                time_field::Event::TypeIntoSegment { kind, ch }
-            )),
-        arb_time_editable_kind().prop_map(|kind| {
-            TimeFieldAction::Send(time_field::Event::TypeBufferCommit { kind })
-        }),
-        arb_time_editable_kind()
-            .prop_map(|kind| TimeFieldAction::Send(time_field::Event::ClearSegment { kind })),
-        Just(TimeFieldAction::Send(time_field::Event::ClearAll)),
-        prop::option::of(arb_time())
-            .prop_map(|value| TimeFieldAction::Send(time_field::Event::SetValue(value))),
-        prop::option::of(arb_time()).prop_map(TimeFieldAction::SetControlledValue),
-        any::<bool>().prop_map(TimeFieldAction::SetDisabled),
-        any::<bool>().prop_map(TimeFieldAction::SetReadonly),
-        any::<bool>().prop_map(TimeFieldAction::SetInvalid),
-    ]
-}
-
-proptest! {
-    #![proptest_config(super::common::proptest_config())]
-
-    #[test]
-    #[ignore = "proptest — nightly extended-proptest job"]
-    fn proptest_date_field_event_sequences_preserve_invariants(
-        actions in prop::collection::vec(arb_date_field_action(), 0..128),
-    ) {
-        let mut service = Service::<date_field::Machine>::new(
-            date_field_props(),
-            &Env::default(),
-            &date_field::Messages::default(),
-        );
-
-        for action in actions {
-            match action {
-                DateFieldAction::Send(event) => {
-                    drop(service.send(event));
-                }
-
-                DateFieldAction::SetControlledValue(value) => {
-                    drop(service.set_props(date_field_props().value(value)));
-                }
-
-                DateFieldAction::SetDisabled(value) => {
-                    drop(service.set_props(date_field_props().disabled(value)));
-                }
-
-                DateFieldAction::SetReadonly(value) => {
-                    drop(service.set_props(date_field_props().readonly(value)));
-                }
-
-                DateFieldAction::SetInvalid(value) => {
-                    drop(service.set_props(date_field_props().invalid(value)));
-                }
-            }
-
-            let ctx = service.context();
-
-            prop_assert_eq!(ctx.ids.id(), "date-field");
-
-            for segment in &ctx.segments {
-                prop_assert!(segment.min <= segment.max);
-
-                if let Some(value) = segment.value {
-                    prop_assert!(value >= segment.min);
-                    prop_assert!(value <= segment.max);
-                }
-            }
-
-            match service.state() {
-                date_field::State::Idle => {}
-
-                date_field::State::Focused(kind) => {
-                    prop_assert!(kind.is_editable());
-                    prop_assert_eq!(ctx.focused_segment, Some(*kind));
-                }
-            }
-
-            if ctx.type_buffer.is_empty()
-                && ctx.pending_controlled_value.is_none()
-                && !ctx.value.is_controlled()
-                && let Some(value) = ctx.value.get()
-            {
-                prop_assert_eq!(
-                    ctx.segments
-                        .iter()
-                        .find(|segment| segment.kind == DateSegmentKind::Year)
-                        .and_then(|segment| segment.value),
-                    Some(value.year())
-                );
-                prop_assert_eq!(
-                    ctx.segments
-                        .iter()
-                        .find(|segment| segment.kind == DateSegmentKind::Month)
-                        .and_then(|segment| segment.value),
-                    Some(i32::from(value.month()))
-                );
-                prop_assert_eq!(
-                    ctx.segments
-                        .iter()
-                        .find(|segment| segment.kind == DateSegmentKind::Day)
-                        .and_then(|segment| segment.value),
-                    Some(i32::from(value.day()))
-                );
-            }
-        }
-    }
-
-    #[test]
-    #[ignore = "proptest — nightly extended-proptest job"]
-    fn proptest_time_field_event_sequences_preserve_invariants(
-        actions in prop::collection::vec(arb_time_field_action(), 0..128),
-    ) {
-        let mut service = Service::<time_field::Machine>::new(
-            time_field_props(),
-            &Env::default(),
-            &time_field::Messages::default(),
-        );
-
-        for action in actions {
-            match action {
-                TimeFieldAction::Send(event) => {
-                    drop(service.send(event));
-                }
-
-                TimeFieldAction::SetControlledValue(value) => {
-                    drop(service.set_props(time_field_props().value(value)));
-                }
-
-                TimeFieldAction::SetDisabled(value) => {
-                    drop(service.set_props(time_field_props().disabled(value)));
-                }
-
-                TimeFieldAction::SetReadonly(value) => {
-                    drop(service.set_props(time_field_props().readonly(value)));
-                }
-
-                TimeFieldAction::SetInvalid(value) => {
-                    drop(service.set_props(time_field_props().invalid(value)));
-                }
-            }
-
-            let ctx = service.context();
-
-            prop_assert_eq!(ctx.ids.id(), "time-field");
-
-            for segment in &ctx.segments {
-                prop_assert!(segment.min <= segment.max);
-
-                if let Some(value) = segment.value {
-                    prop_assert!(value >= segment.min);
-                    prop_assert!(value <= segment.max);
-                }
-            }
-
-            if matches!(ctx.hour_cycle, HourCycle::H23 | HourCycle::H24) {
-                prop_assert!(
-                    !ctx.segments
-                        .iter()
-                        .any(|segment| segment.kind == DateSegmentKind::DayPeriod)
-                );
-            }
-
-            match service.state() {
-                time_field::State::Idle => {}
-
-                time_field::State::Focused(kind) => {
-                    prop_assert!(kind.is_editable());
-                    prop_assert_eq!(ctx.focused_segment, Some(*kind));
-                }
-            }
-
-            if ctx.type_buffer.is_empty()
-                && !ctx.value.is_controlled()
-                && let Some(value) = ctx.value.get()
-            {
-                prop_assert_eq!(
-                    ctx.segments
-                        .iter()
-                        .find(|segment| segment.kind == DateSegmentKind::Hour)
-                        .and_then(|segment| segment.value),
-                    Some(match ctx.hour_cycle {
-                        HourCycle::H11 => i32::from(value.hour() % 12),
-                        HourCycle::H12 => i32::from(if value.hour() % 12 == 0 { 12 } else { value.hour() % 12 }),
-                        HourCycle::H23 => i32::from(value.hour()),
-                        HourCycle::H24 => i32::from(if value.hour() == 0 { 24 } else { value.hour() }),
-                    })
-                );
-            }
-        }
-    }
-}
-
-// ────────────────────────────────────────────────────────────────────
-// Calendar
-// ────────────────────────────────────────────────────────────────────
-
-#[derive(Clone, Debug)]
-enum CalendarAction {
-    Send(calendar::Event),
-    SetDisabled(bool),
-    SetReadonly(bool),
-    SetMin(Option<CalendarDate>),
-    SetMax(Option<CalendarDate>),
-}
-
-fn calendar_props() -> calendar::Props {
-    calendar::Props::new().id("cal").today(date(2024, 1, 15))
-}
-
-fn calendar_props_multi() -> calendar::Props {
-    calendar_props().selection_mode(SelectionMode::Multiple)
-}
-
-fn arb_calendar_date() -> impl Strategy<Value = CalendarDate> {
-    (1900i32..=2100, 1u8..=12, 1u8..=28).prop_map(|(y, m, d)| date(y, m, d))
-}
-
-fn arb_calendar_key() -> impl Strategy<Value = KeyboardKey> {
-    prop_oneof![
-        Just(KeyboardKey::ArrowLeft),
-        Just(KeyboardKey::ArrowRight),
-        Just(KeyboardKey::ArrowUp),
-        Just(KeyboardKey::ArrowDown),
-        Just(KeyboardKey::Home),
-        Just(KeyboardKey::End),
-        Just(KeyboardKey::PageUp),
-        Just(KeyboardKey::PageDown),
-        Just(KeyboardKey::Enter),
-        Just(KeyboardKey::Space),
-    ]
-}
-
-fn arb_calendar_event() -> impl Strategy<Value = calendar::Event> {
-    prop_oneof![
-        arb_calendar_date().prop_map(|date| calendar::Event::FocusDate { date }),
-        arb_calendar_date().prop_map(|date| calendar::Event::SelectDate { date }),
-        arb_calendar_date().prop_map(|date| calendar::Event::ToggleDate { date }),
-        Just(calendar::Event::NextMonth),
-        Just(calendar::Event::PrevMonth),
-        Just(calendar::Event::NextYear),
-        Just(calendar::Event::PrevYear),
-        (1u8..=12).prop_map(|month| calendar::Event::SetMonth { month }),
-        (1900i32..=2100).prop_map(|year| calendar::Event::SetYear { year }),
-        Just(calendar::Event::FocusIn),
-        Just(calendar::Event::FocusOut),
-        (arb_calendar_key(), any::<bool>())
-            .prop_map(|(key, shift)| calendar::Event::KeyDown { key, shift }),
-    ]
-}
-
-fn arb_calendar_action() -> impl Strategy<Value = CalendarAction> {
-    prop_oneof![
-        arb_calendar_event().prop_map(CalendarAction::Send),
-        any::<bool>().prop_map(CalendarAction::SetDisabled),
-        any::<bool>().prop_map(CalendarAction::SetReadonly),
-        prop::option::of(arb_calendar_date()).prop_map(CalendarAction::SetMin),
-        prop::option::of(arb_calendar_date()).prop_map(CalendarAction::SetMax),
-    ]
-}
-
-fn apply_calendar_action(
-    service: &mut Service<calendar::Machine>,
-    action: CalendarAction,
-    base_props: &calendar::Props,
-) {
-    match action {
-        CalendarAction::Send(event) => {
-            drop(service.send(event));
-        }
-
-        CalendarAction::SetDisabled(value) => {
-            drop(service.set_props(base_props.clone().disabled(value)));
-        }
-
-        CalendarAction::SetReadonly(value) => {
-            drop(service.set_props(base_props.clone().readonly(value)));
-        }
-
-        CalendarAction::SetMin(value) => {
-            drop(service.set_props(base_props.clone().min(value)));
-        }
-
-        CalendarAction::SetMax(value) => {
-            drop(service.set_props(base_props.clone().max(value)));
-        }
-    }
-}
-
-fn assert_calendar_invariants(service: &Service<calendar::Machine>) {
-    let ctx = service.context();
-
-    assert!(ctx.visible_month >= 1, "visible_month must be 1-based");
-    assert!(ctx.visible_month <= 12, "visible_month must be 1..=12");
-    assert!(ctx.visible_months >= 1, "visible_months must be >= 1");
-    assert!(
-        matches!(
-            ctx.first_day_of_week,
-            Weekday::Sunday
-                | Weekday::Monday
-                | Weekday::Tuesday
-                | Weekday::Wednesday
-                | Weekday::Thursday
-                | Weekday::Friday
-                | Weekday::Saturday,
-        ),
-        "first_day_of_week must be a valid weekday",
-    );
-
-    if let Some(min) = &ctx.min {
-        assert!(
-            !matches!(ctx.focused_date.compare(min), core::cmp::Ordering::Less),
-            "focused_date must be >= min",
-        );
-    }
-
-    if let Some(max) = &ctx.max {
-        assert!(
-            !matches!(ctx.focused_date.compare(max), core::cmp::Ordering::Greater),
-            "focused_date must be <= max",
-        );
-    }
-
-    if let Some(cap) = ctx.max_selected {
-        assert!(
-            ctx.selected_dates.get().len() <= cap,
-            "selected_dates exceeded max_selected cap",
-        );
-    }
-
-    let weeks = ctx.weeks();
-
-    if !weeks.is_empty() {
-        assert_eq!(weeks.len(), 6, "grid must always render 6 weeks");
-        for row in &weeks {
-            assert_eq!(row.len(), 7, "every row must contain 7 days");
-        }
-    }
-
-    assert!(matches!(
-        service.state(),
-        calendar::State::Idle | calendar::State::Focused,
-    ));
-}
-
-proptest! {
-    #![proptest_config(super::common::proptest_config())]
-
-    #[test]
-    #[ignore = "proptest — nightly extended-proptest job"]
-    fn proptest_calendar_single_select_preserves_invariants(
-        actions in prop::collection::vec(arb_calendar_action(), 0..64),
-    ) {
-        let base = calendar_props();
-
-        let mut service = Service::<calendar::Machine>::new(
-            base.clone(),
-            &Env::default(),
-            &calendar::Messages::default(),
-        );
-
-        for action in actions {
-            apply_calendar_action(&mut service, action, &base);
-
-            assert_calendar_invariants(&service);
-        }
-    }
-
-    #[test]
-    #[ignore = "proptest — nightly extended-proptest job"]
-    fn proptest_calendar_multi_select_preserves_invariants(
-        actions in prop::collection::vec(arb_calendar_action(), 0..64),
-        max_selected in prop::option::of(1usize..=8),
-    ) {
-        let base = calendar_props_multi().max_selected(max_selected);
-
-        let mut service = Service::<calendar::Machine>::new(
-            base.clone(),
-            &Env::default(),
-            &calendar::Messages::default(),
-        );
-
-        for action in actions {
-            apply_calendar_action(&mut service, action, &base);
-
-            assert_calendar_invariants(&service);
-        }
-    }
-
-    #[test]
-    #[ignore = "proptest — nightly extended-proptest job"]
-    fn proptest_calendar_page_behavior_advances_one_or_visible_months(
-        visible in 1usize..=4,
-        behavior in prop_oneof![Just(PageBehavior::Visible), Just(PageBehavior::Single)],
-        steps in 0u8..=20,
-    ) {
-        let base = calendar_props().visible_months(visible).page_behavior(behavior);
-
-        let mut service = Service::<calendar::Machine>::new(
-            base.clone(),
-            &Env::default(),
-            &calendar::Messages::default(),
-        );
-
-        let start_month = i32::from(service.context().visible_month);
-        let start_year = service.context().visible_year;
-
-        for _ in 0..steps {
-            drop(service.send(calendar::Event::NextMonth));
-        }
-
-        let step = match behavior {
-            PageBehavior::Visible => i32::try_from(visible).unwrap_or(1),
-            PageBehavior::Single => 1,
-        };
-
-        let expected_total = start_month - 1 + step * i32::from(steps);
-        let expected_month = expected_total.rem_euclid(12) + 1;
-        let expected_year = start_year + expected_total.div_euclid(12);
-
-        prop_assert_eq!(service.context().visible_month, expected_month as u8);
-        prop_assert_eq!(service.context().visible_year, expected_year);
-    }
-}
-
-// ────────────────────────────────────────────────────────────────────
-// RangeCalendar
-// ────────────────────────────────────────────────────────────────────
-
-#[derive(Clone, Debug)]
-enum RangeCalendarAction {
-    Send(range_calendar::Event),
-    SetDisabled(bool),
-    SetReadonly(bool),
-    SetMin(Option<CalendarDate>),
-    SetMax(Option<CalendarDate>),
-}
-
-fn range_calendar_props() -> range_calendar::Props {
-    range_calendar::Props::new()
-        .id("range-calendar")
-        .today(date(2024, 1, 15))
-        .max_range_days(Some(14))
-}
-
-fn arb_range_calendar_key() -> impl Strategy<Value = KeyboardKey> {
-    prop_oneof![
-        Just(KeyboardKey::ArrowLeft),
-        Just(KeyboardKey::ArrowRight),
-        Just(KeyboardKey::ArrowUp),
-        Just(KeyboardKey::ArrowDown),
-        Just(KeyboardKey::Home),
-        Just(KeyboardKey::End),
-        Just(KeyboardKey::PageUp),
-        Just(KeyboardKey::PageDown),
-        Just(KeyboardKey::Enter),
-        Just(KeyboardKey::Space),
-    ]
-}
-
-fn arb_range_calendar_event() -> impl Strategy<Value = range_calendar::Event> {
-    prop_oneof![
-        arb_calendar_date().prop_map(|date| range_calendar::Event::FocusDate { date }),
-        arb_calendar_date().prop_map(|date| range_calendar::Event::SelectDate { date }),
-        arb_calendar_date().prop_map(|date| range_calendar::Event::HoverDate { date }),
-        Just(range_calendar::Event::HoverEnd),
-        Just(range_calendar::Event::NextMonth),
-        Just(range_calendar::Event::PrevMonth),
-        Just(range_calendar::Event::NextYear),
-        Just(range_calendar::Event::PrevYear),
-        (1u8..=12).prop_map(|month| range_calendar::Event::SetMonth { month }),
-        (1900i32..=2100).prop_map(|year| range_calendar::Event::SetYear { year }),
-        Just(range_calendar::Event::FocusIn),
-        Just(range_calendar::Event::FocusOut),
-        (arb_range_calendar_key(), any::<bool>())
-            .prop_map(|(key, shift)| range_calendar::Event::KeyDown { key, shift }),
-    ]
-}
-
-fn arb_range_calendar_action() -> impl Strategy<Value = RangeCalendarAction> {
-    prop_oneof![
-        arb_range_calendar_event().prop_map(RangeCalendarAction::Send),
-        any::<bool>().prop_map(RangeCalendarAction::SetDisabled),
-        any::<bool>().prop_map(RangeCalendarAction::SetReadonly),
-        prop::option::of(arb_calendar_date()).prop_map(RangeCalendarAction::SetMin),
-        prop::option::of(arb_calendar_date()).prop_map(RangeCalendarAction::SetMax),
-    ]
-}
-
-fn apply_range_calendar_action(
-    service: &mut Service<range_calendar::Machine>,
-    action: RangeCalendarAction,
-    base_props: &range_calendar::Props,
-) {
-    match action {
-        RangeCalendarAction::Send(event) => {
-            drop(service.send(event));
-        }
-
-        RangeCalendarAction::SetDisabled(value) => {
-            drop(service.set_props(base_props.clone().disabled(value)));
-        }
-
-        RangeCalendarAction::SetReadonly(value) => {
-            drop(service.set_props(base_props.clone().readonly(value)));
-        }
-
-        RangeCalendarAction::SetMin(value) => {
-            drop(service.set_props(base_props.clone().min(value)));
-        }
-
-        RangeCalendarAction::SetMax(value) => {
-            drop(service.set_props(base_props.clone().max(value)));
-        }
-    }
-}
-
-fn assert_range_calendar_invariants(service: &Service<range_calendar::Machine>) {
-    let ctx = service.context();
-
-    assert!(ctx.visible_month >= 1, "visible_month must be 1-based");
-    assert!(ctx.visible_month <= 12, "visible_month must be 1..=12");
-    assert!(ctx.visible_months >= 1, "visible_months must be >= 1");
-
-    if let Some(min) = &ctx.min {
-        assert!(
-            !matches!(ctx.focused_date.compare(min), core::cmp::Ordering::Less),
-            "focused_date must be >= min",
-        );
-    }
-
-    if let Some(max) = &ctx.max {
-        assert!(
-            !matches!(ctx.focused_date.compare(max), core::cmp::Ordering::Greater),
-            "focused_date must be <= max",
-        );
-    }
-
-    if let Some(range) = ctx.value.get() {
-        assert!(
-            !matches!(
-                range.start.compare(&range.end),
-                core::cmp::Ordering::Greater,
-            ),
-            "range must be normalized",
-        );
-        assert!(
-            ctx.range_is_allowed(range),
-            "completed range must satisfy span constraints",
-        );
-    }
-
-    if ctx.anchor_date.is_none() {
-        assert!(
-            ctx.hovering_date.is_none(),
-            "hovering_date must not outlive a pending anchor",
-        );
-    }
-
-    let weeks = ctx.weeks();
-
-    if !weeks.is_empty() {
-        assert_eq!(weeks.len(), 6, "grid must always render 6 weeks");
-
-        for row in &weeks {
-            assert_eq!(row.len(), 7, "every row must contain 7 days");
-        }
-    }
-
-    assert!(matches!(
-        service.state(),
-        range_calendar::State::Idle | range_calendar::State::Focused,
-    ));
-}
-
-proptest! {
-    #![proptest_config(super::common::proptest_config())]
-
-    #[test]
-    #[ignore = "proptest — nightly extended-proptest job"]
-    fn proptest_range_calendar_sequences_preserve_invariants(
-        actions in prop::collection::vec(arb_range_calendar_action(), 0..64),
-        allow_single in any::<bool>(),
-        min_range_days in prop::option::of(1u32..=5),
-        max_range_days in prop::option::of(6u32..=21),
-    ) {
-        let base = range_calendar_props()
-            .allow_single_date_range(allow_single)
-            .min_range_days(min_range_days)
-            .max_range_days(max_range_days);
-
-        let mut service = Service::<range_calendar::Machine>::new(
-            base.clone(),
-            &Env::default(),
-            &range_calendar::Messages::default(),
-        );
-
-        for action in actions {
-            apply_range_calendar_action(&mut service, action, &base);
-
-            assert_range_calendar_invariants(&service);
-        }
-    }
-
-    #[test]
-    #[ignore = "proptest — nightly extended-proptest job"]
-    fn proptest_range_calendar_accepts_only_allowed_completed_ranges(
-        first in arb_calendar_date(),
-        second in arb_calendar_date(),
-        max_days in 1u32..=20,
-    ) {
-        let base = range_calendar_props().max_range_days(Some(max_days));
-        let mut service = Service::<range_calendar::Machine>::new(
-            base,
-            &Env::default(),
-            &range_calendar::Messages::default(),
-        );
-
-        drop(service.send(range_calendar::Event::SelectDate { date: first.clone() }));
-        drop(service.send(range_calendar::Event::SelectDate { date: second.clone() }));
-
-        if let Some(expected) = DateRange::normalized(first, second)
-            && expected
-                .start
-                .days_until(&expected.end)
-                .ok()
-                .and_then(|days| u32::try_from(days).ok())
-                .and_then(|days| days.checked_add(1))
-                .is_some_and(|days| days <= max_days)
-        {
-            prop_assert_eq!(service.context().value.get(), &Some(expected));
-        } else {
-            prop_assert!(service.context().value.get().is_none());
-            prop_assert!(service.context().anchor_date.is_some());
-        }
-    }
-}
-
-// ────────────────────────────────────────────────────────────────────
-// DatePicker
-// ────────────────────────────────────────────────────────────────────
-
-#[derive(Clone, Debug)]
-enum DatePickerAction {
-    Send(date_picker::Event),
-    SetControlledValue(Option<CalendarDate>),
-    SetControlledOpen(Option<bool>),
-    SetDisabled(bool),
-}
-
-fn date_picker_props() -> date_picker::Props {
-    date_picker::Props {
-        id: "date-picker".to_string(),
-        label: "Date".to_string(),
-        ..date_picker::Props::default()
-    }
-}
-
-fn arb_date_picker_event() -> impl Strategy<Value = date_picker::Event> {
-    prop_oneof![
-        Just(date_picker::Event::Open),
-        Just(date_picker::Event::Close),
-        Just(date_picker::Event::Toggle),
-        arb_date().prop_map(|date| date_picker::Event::SelectDate { date }),
-        "[0-9/]{0,12}".prop_map(|value| date_picker::Event::InputChange { value }),
-        Just(date_picker::Event::FocusIn),
-        Just(date_picker::Event::FocusOut),
-        Just(date_picker::Event::KeyDown {
-            key: KeyboardKey::Escape,
-        }),
-        Just(date_picker::Event::KeyDown {
-            key: KeyboardKey::ArrowDown,
-        }),
-    ]
-}
-
-fn arb_date_picker_action() -> impl Strategy<Value = DatePickerAction> {
-    prop_oneof![
-        arb_date_picker_event().prop_map(DatePickerAction::Send),
-        prop_oneof![Just(None), arb_date().prop_map(Some)]
-            .prop_map(DatePickerAction::SetControlledValue),
-        prop_oneof![Just(None), Just(Some(true)), Just(Some(false))]
-            .prop_map(DatePickerAction::SetControlledOpen),
-        any::<bool>().prop_map(DatePickerAction::SetDisabled),
-    ]
-}
-
-proptest! {
-    #![proptest_config(super::common::proptest_config())]
-
-    /// Random event and controlled-prop sequences keep the picker's invariants:
-    /// the component ID stays stable, the state is always one of its two
-    /// variants, and connecting every anatomy part never panics (exercising
-    /// `parse_date` on arbitrary text in particular).
-    #[test]
-    #[ignore = "proptest — nightly extended-proptest job"]
-    fn proptest_date_picker_sequences_preserve_invariants(
-        actions in prop::collection::vec(arb_date_picker_action(), 0..128),
-    ) {
-        let mut service = Service::<date_picker::Machine>::new(
-            date_picker_props(),
-            &Env::default(),
-            &date_picker::Messages::default(),
-        );
-
-        for action in actions {
-            match action {
-                DatePickerAction::Send(event) => {
-                    drop(service.send(event));
-                }
-
-                DatePickerAction::SetControlledValue(value) => {
-                    drop(service.set_props(date_picker::Props {
-                        value: Some(value),
-                        ..date_picker_props()
-                    }));
-                }
-
-                DatePickerAction::SetControlledOpen(open) => {
-                    drop(service.set_props(date_picker::Props {
-                        open,
-                        ..date_picker_props()
-                    }));
-                }
-
-                DatePickerAction::SetDisabled(disabled) => {
-                    drop(service.set_props(date_picker::Props {
-                        disabled,
-                        ..date_picker_props()
-                    }));
-                }
-            }
-
-            prop_assert_eq!(service.context().ids.id(), "date-picker");
-            prop_assert!(matches!(
-                service.state(),
-                date_picker::State::Open | date_picker::State::Closed,
-            ));
-
-            let send = |_event: date_picker::Event| {};
-
-            let api = service.connect(&send);
-
-            for part in date_picker::Part::all() {
-                drop(api.part_attrs(part));
-            }
-        }
-    }
-}
-
-// ────────────────────────────────────────────────────────────────────
-// DateTimePicker
-// ────────────────────────────────────────────────────────────────────
-
-fn arb_datetime() -> impl Strategy<Value = CalendarDateTime> {
-    (arb_date(), arb_time()).prop_map(|(date, time)| CalendarDateTime::new(date, time))
-}
-
-#[derive(Clone, Debug)]
-enum DateTimePickerAction {
-    Send(date_time_picker::Event),
-    SetControlledValue(Option<CalendarDateTime>),
-    SetDisabled(bool),
-    SetReadonly(bool),
-}
-
-fn date_time_picker_props() -> date_time_picker::Props {
-    date_time_picker::Props {
-        id: "date-time-picker".to_string(),
-        label: "Appointment".to_string(),
-        ..date_time_picker::Props::default()
-    }
-}
-
-fn arb_date_time_picker_segment() -> impl Strategy<Value = DateSegmentKind> {
-    prop_oneof![
-        Just(DateSegmentKind::Year),
-        Just(DateSegmentKind::Month),
-        Just(DateSegmentKind::Day),
-        Just(DateSegmentKind::Hour),
-        Just(DateSegmentKind::Minute),
-        Just(DateSegmentKind::DayPeriod),
-    ]
-}
-
-fn arb_date_time_picker_event() -> impl Strategy<Value = date_time_picker::Event> {
-    prop_oneof![
-        Just(date_time_picker::Event::Open),
-        Just(date_time_picker::Event::Close),
-        Just(date_time_picker::Event::Toggle),
-        Just(date_time_picker::Event::FocusIn),
-        Just(date_time_picker::Event::FocusOut),
-        Just(date_time_picker::Event::FocusNextSegment),
-        Just(date_time_picker::Event::FocusPrevSegment),
-        Just(date_time_picker::Event::ClearAll),
-        arb_date().prop_map(date_time_picker::Event::CalendarSelectDate),
-        arb_date_time_picker_segment().prop_map(date_time_picker::Event::FocusSegment),
-        arb_date_time_picker_segment()
-            .prop_map(|segment| date_time_picker::Event::IncrementSegment { segment }),
-        arb_date_time_picker_segment()
-            .prop_map(|segment| date_time_picker::Event::DecrementSegment { segment }),
-        (arb_date_time_picker_segment(), 0i32..=60)
-            .prop_map(|(segment, value)| date_time_picker::Event::SegmentChange { segment, value }),
-        (arb_date_time_picker_segment(), prop::char::range('0', '9'))
-            .prop_map(|(segment, ch)| date_time_picker::Event::TypeIntoSegment { segment, ch }),
-        arb_date_time_picker_segment()
-            .prop_map(|segment| date_time_picker::Event::ClearSegment { segment }),
-        prop_oneof![Just(None), arb_datetime().prop_map(Some)]
-            .prop_map(date_time_picker::Event::ValueCommit),
-        Just(date_time_picker::Event::KeyDown {
-            key: KeyboardKey::Escape,
-        }),
-        Just(date_time_picker::Event::KeyDown {
-            key: KeyboardKey::ArrowDown,
-        }),
-    ]
-}
-
-fn arb_date_time_picker_action() -> impl Strategy<Value = DateTimePickerAction> {
-    prop_oneof![
-        arb_date_time_picker_event().prop_map(DateTimePickerAction::Send),
-        prop_oneof![Just(None), arb_datetime().prop_map(Some)]
-            .prop_map(DateTimePickerAction::SetControlledValue),
-        any::<bool>().prop_map(DateTimePickerAction::SetDisabled),
-        any::<bool>().prop_map(DateTimePickerAction::SetReadonly),
-    ]
-}
-
-proptest! {
-    #![proptest_config(super::common::proptest_config())]
-
-    /// Random event and controlled-prop sequences keep the picker's invariants:
-    /// the component ID stays stable, the state is always one of its three
-    /// variants, every segment's value stays within its `[min, max]` range, a
-    /// 24-hour cycle never grows a day-period segment, the focused segment is
-    /// always editable, a committed value never falls outside `[min, max]`, and
-    /// connecting every anatomy part never panics.
-    #[test]
-    #[ignore = "proptest — nightly extended-proptest job"]
-    fn proptest_date_time_picker_sequences_preserve_invariants(
-        actions in prop::collection::vec(arb_date_time_picker_action(), 0..128),
-    ) {
-        let mut service = Service::<date_time_picker::Machine>::new(
-            date_time_picker_props(),
-            &Env::default(),
-            &date_time_picker::Messages::default(),
-        );
-
-        for action in actions {
-            match action {
-                DateTimePickerAction::Send(event) => {
-                    drop(service.send(event));
-                }
-
-                DateTimePickerAction::SetControlledValue(value) => {
-                    drop(service.set_props(date_time_picker::Props {
-                        value: Some(value),
-                        ..date_time_picker_props()
-                    }));
-                }
-
-                DateTimePickerAction::SetDisabled(disabled) => {
-                    drop(service.set_props(date_time_picker::Props {
-                        disabled,
-                        ..date_time_picker_props()
-                    }));
-                }
-
-                DateTimePickerAction::SetReadonly(readonly) => {
-                    drop(service.set_props(date_time_picker::Props {
-                        readonly,
-                        ..date_time_picker_props()
-                    }));
-                }
-            }
-
-            let ctx = service.context();
-
-            prop_assert_eq!(ctx.ids.id(), "date-time-picker");
-            prop_assert!(matches!(
-                service.state(),
-                date_time_picker::State::Idle
-                    | date_time_picker::State::Focused
-                    | date_time_picker::State::Open,
-            ));
-
-            for segment in ctx.all_segments() {
-                prop_assert!(segment.min <= segment.max);
-
-                if let Some(value) = segment.value {
-                    prop_assert!(value >= segment.min);
-                    prop_assert!(value <= segment.max);
-                }
-            }
-
-            if matches!(ctx.hour_cycle, HourCycle::H23 | HourCycle::H24) {
-                prop_assert!(
-                    !ctx.all_segments()
-                        .any(|segment| segment.kind == DateSegmentKind::DayPeriod)
-                );
-            }
-
-            if let Some(kind) = ctx.focused_segment {
-                prop_assert!(
-                    ctx.segment(kind).is_some_and(|segment| segment.is_editable)
-                );
-            }
-
-            let send = |_event: date_time_picker::Event| {};
-
-            let api = service.connect(&send);
-
-            for part in date_time_picker::Part::all() {
-                drop(api.part_attrs(part));
-            }
-        }
-    }
-}
-
-// ── DateRangeField ───────────────────────────────────────────────────────
-
-fn arb_date_range_field_event() -> impl Strategy<Value = date_range_field::Event> {
-    prop_oneof![
-        Just(date_range_field::Event::FocusStart),
-        Just(date_range_field::Event::FocusEnd),
-        Just(date_range_field::Event::BlurAll),
-        proptest::option::of(arb_date()).prop_map(date_range_field::Event::StartValueChange),
-        proptest::option::of(arb_date()).prop_map(date_range_field::Event::EndValueChange),
-        proptest::option::of((arb_date(), arb_date()).prop_map(|(first, second)| {
-            DateRange::normalized(first, second).expect("comparable generated dates")
-        }))
-        .prop_map(date_range_field::Event::SetRange),
-    ]
-}
-
-fn assert_date_range_field_invariants(service: &Service<date_range_field::Machine>) {
-    let ctx = service.context();
-
-    // The derived complete range is `Some` exactly when both fields are set,
-    // and is always normalized so `start <= end`.
-    if let Some(range) = ctx.value.get() {
-        assert!(
-            ctx.start_date.is_some() && ctx.end_date.is_some(),
-            "complete range requires both fields set"
-        );
-        assert!(
-            matches!(
-                range.start.compare_within_calendar(&range.end),
-                Some(core::cmp::Ordering::Less | core::cmp::Ordering::Equal)
-            ),
-            "stored range must be normalized so start <= end"
-        );
-        assert_eq!(ctx.start_date.as_ref(), Some(&range.start));
-        assert_eq!(ctx.end_date.as_ref(), Some(&range.end));
-    } else {
-        assert!(
-            ctx.start_date.is_none() || ctx.end_date.is_none(),
-            "an empty range must have at least one unset field"
-        );
-    }
-
-    // Every part produces attributes without panicking.
-    let send = |_event: date_range_field::Event| {};
-
-    let api = service.connect(&send);
-
-    for part in date_range_field::Part::all() {
-        drop(api.part_attrs(part));
-    }
-
-    // Child field props mirror the tracked per-field values.
-    assert_eq!(api.start_field_props().value, Some(ctx.start_date.clone()));
-    assert_eq!(api.end_field_props().value, Some(ctx.end_date.clone()));
-}
-
-proptest! {
-    #![proptest_config(super::common::proptest_config())]
-
-    #[test]
-    fn date_range_field_maintains_normalized_range(
-        events in prop::collection::vec(arb_date_range_field_event(), 0..24),
-    ) {
-        let mut service = Service::<date_range_field::Machine>::new(
-            date_range_field::Props::new().id("range"),
-            &Env::default(),
-            &date_range_field::Messages::default(),
-        );
-
-        for event in events {
-            drop(service.send(event));
-
-            assert_date_range_field_invariants(&service);
-        }
-    }
-
-    #[test]
-    fn date_range_field_disabled_ignores_events(
-        events in prop::collection::vec(arb_date_range_field_event(), 0..24),
-    ) {
-        let mut service = Service::<date_range_field::Machine>::new(
-            date_range_field::Props::new().id("range").disabled(true),
-            &Env::default(),
-            &date_range_field::Messages::default(),
-        );
-
-        for event in events {
-            drop(service.send(event));
-        }
-
-        prop_assert_eq!(service.state(), &date_range_field::State::Idle);
-        prop_assert_eq!(service.context().value.get().as_ref(), None);
-    }
-}
-
-// ── DateRangePicker ────────────────────────────────────────────────────────
-
-#[derive(Clone, Debug)]
-enum DateRangePickerAction {
-    Send(date_range_picker::Event),
-    SetControlledValue(Option<DateRange>),
-    SetDisabled(bool),
-}
-
-fn date_range_picker_presets() -> Vec<date_range_picker::Preset> {
-    vec![
-        date_range_picker::Preset::new(
-            "Last 7 days",
-            DateRange::new(date(2025, 5, 26), date(2025, 6, 1)).expect("ordered range"),
-        ),
-        date_range_picker::Preset::new(
-            "Last 30 days",
-            DateRange::new(date(2025, 5, 3), date(2025, 6, 1)).expect("ordered range"),
-        ),
-    ]
-}
-
-fn date_range_picker_props() -> date_range_picker::Props {
-    date_range_picker::Props {
-        id: "range-picker".to_string(),
-        presets: date_range_picker_presets(),
-        ..date_range_picker::Props::default()
-    }
-}
-
-fn arb_date_range_picker_event() -> impl Strategy<Value = date_range_picker::Event> {
-    prop_oneof![
-        Just(date_range_picker::Event::Open),
-        Just(date_range_picker::Event::Close),
-        Just(date_range_picker::Event::Toggle),
-        (arb_date(), arb_date())
-            .prop_map(|(first, second)| {
-                DateRange::normalized(first, second).expect("comparable generated dates")
-            })
-            .prop_map(|range| date_range_picker::Event::SelectRangeComplete { range }),
-        proptest::option::of(arb_date()).prop_map(date_range_picker::Event::StartValueChange),
-        proptest::option::of(arb_date()).prop_map(date_range_picker::Event::EndValueChange),
-        (0usize..=3).prop_map(|index| date_range_picker::Event::SelectPreset { index }),
-        Just(date_range_picker::Event::Clear),
-        Just(date_range_picker::Event::FocusIn),
-        Just(date_range_picker::Event::FocusOut),
-        Just(date_range_picker::Event::KeyDown {
-            key: KeyboardKey::Escape,
-        }),
-        Just(date_range_picker::Event::KeyDown {
-            key: KeyboardKey::ArrowDown,
-        }),
-    ]
-}
-
-fn arb_date_range_picker_action() -> impl Strategy<Value = DateRangePickerAction> {
-    prop_oneof![
-        arb_date_range_picker_event().prop_map(DateRangePickerAction::Send),
-        prop_oneof![
-            Just(None),
-            (arb_date(), arb_date())
-                .prop_map(|(first, second)| DateRange::normalized(first, second))
-        ]
-        .prop_map(DateRangePickerAction::SetControlledValue),
-        any::<bool>().prop_map(DateRangePickerAction::SetDisabled),
-    ]
-}
-
-proptest! {
-    #![proptest_config(super::common::proptest_config())]
-
-    /// Random event and controlled-prop sequences keep the picker's invariants:
-    /// the ID stays stable, the state is one of its two variants, the canonical
-    /// range stays normalized and consistent with the per-field values, and
-    /// connecting every anatomy part never panics.
-    #[test]
-    #[ignore = "proptest — nightly extended-proptest job"]
-    fn proptest_date_range_picker_sequences_preserve_invariants(
-        actions in prop::collection::vec(arb_date_range_picker_action(), 0..128),
-    ) {
-        let mut service = Service::<date_range_picker::Machine>::new(
-            date_range_picker_props(),
-            &Env::default(),
-            &date_range_picker::Messages::default(),
-        );
-
-        for action in actions {
-            match action {
-                DateRangePickerAction::Send(event) => {
-                    drop(service.send(event));
-                }
-
-                DateRangePickerAction::SetControlledValue(value) => {
-                    drop(service.set_props(date_range_picker::Props {
-                        value: Some(value),
-                        ..date_range_picker_props()
-                    }));
-                }
-
-                DateRangePickerAction::SetDisabled(disabled) => {
-                    drop(service.set_props(date_range_picker::Props {
-                        disabled,
-                        ..date_range_picker_props()
-                    }));
-                }
-            }
-
-            let ctx = service.context();
-
-            prop_assert_eq!(ctx.ids.id(), "range-picker");
-            prop_assert!(matches!(
-                service.state(),
-                date_range_picker::State::Open | date_range_picker::State::Closed,
-            ));
-
-            // A selected range is always normalized so `start <= end` (in both
-            // modes). In *uncontrolled* mode the canonical range additionally
-            // mirrors the per-field values — it is `Some` exactly when both
-            // fields are set. In controlled mode `get()` returns the parent's
-            // override, which is intentionally decoupled from local field edits
-            // (the parent reconciles via the next `SyncProps`), so the mirror
-            // check does not apply.
-            if let Some(range) = ctx.value.get() {
-                prop_assert!(matches!(
-                    range.start.compare_within_calendar(&range.end),
-                    Some(core::cmp::Ordering::Less | core::cmp::Ordering::Equal)
-                ));
-
-                if !ctx.value.is_controlled() {
-                    prop_assert!(ctx.start_date.is_some() && ctx.end_date.is_some());
-                    prop_assert_eq!(ctx.start_date.as_ref(), Some(&range.start));
-                    prop_assert_eq!(ctx.end_date.as_ref(), Some(&range.end));
-                }
-            } else if !ctx.value.is_controlled() {
-                prop_assert!(ctx.start_date.is_none() || ctx.end_date.is_none());
-            }
-
-            let send = |_event: date_range_picker::Event| {};
-
-            let api = service.connect(&send);
-
-            for part in date_range_picker::Part::all() {
-                drop(api.part_attrs(part));
-            }
-        }
-    }
-}
