+++ crates/ars-components/tests/spec_conformance/date_time/time_field.rs
@@ -1,0 +1,25 @@
+use ars_components::date_time::{date_field::DateSegmentKind, time_field};
+
+use super::assert_anatomy;
+
+#[test]
+fn time_field_anatomy_matches_spec() {
+    assert_anatomy(
+        "time-field",
+        &[
+            (time_field::Part::Root, "root"),
+            (time_field::Part::Label, "label"),
+            (time_field::Part::FieldGroup, "field-group"),
+            (
+                time_field::Part::Segment {
+                    kind: DateSegmentKind::Year,
+                },
+                "segment",
+            ),
+            (time_field::Part::Literal { index: 0 }, "literal"),
+            (time_field::Part::Description, "description"),
+            (time_field::Part::ErrorMessage, "error-message"),
+            (time_field::Part::HiddenInput, "hidden-input"),
+        ],
+    );
+}
